move symbol logic to separate file, resolve imports
