+++ bi/src/main.rs
@@ -8,16 +8,22 @@ mod constraints;
 mod proto_trace;
 mod signal_trace;
 
+use baa::BitVecOps;
+use clap::ColorChoice;
+use clap::Parser;
+use clap_verbosity_flag::Verbosity;
+use clap_verbosity_flag::WarnLevel;
+use protocols::frontend;
+use protocols::frontend::design::Design;
+use protocols::frontend::design::find_designs;
+use protocols::frontend::diagnostic::DiagnosticHandler;
+use protocols::frontend::diagnostic::Level;
+use rustc_hash::FxHashMap;
+use rustc_hash::FxHashSet;
+
 use crate::bi::*;
 use crate::proto_trace::*;
 use crate::signal_trace::*;
-use baa::BitVecOps;
-use clap::{ColorChoice, Parser};
-use clap_verbosity_flag::{Verbosity, WarnLevel};
-use protocols::frontend;
-use protocols::frontend::design::{Design, find_designs};
-use protocols::frontend::diagnostic::{DiagnosticHandler, Level};
-use rustc_hash::{FxHashMap, FxHashSet};
 
 /// Args for the monitor CLI
 #[derive(Parser, Debug)]
