+++ monitor/src/scheduler.rs
@@ -1,21 +1,30 @@
-use crate::interpreter::serialize_monitor_args_mapping;
-use crate::{
-    global_context::GlobalContext,
-    interpreter::Interpreter,
-    queue::*,
-    signal_trace::{SignalTrace, WaveSignalTrace},
-    thread::Thread,
-    types::{
-        AugmentedProtocolApplication, AugmentedTrace, CycleResult, LoopArgState, SchedulerError,
-        ThreadResult,
-    },
-};
 use anyhow::anyhow;
-use baa::{BitVecOps, BitVecValue};
-use protocols::errors::{EvaluationError, ExecutionError};
-use protocols::frontend::ast::{Expr, Protocol, Stmt, StmtId, SymbolId, SymbolTable};
+use baa::BitVecOps;
+use baa::BitVecValue;
+use protocols::errors::EvaluationError;
+use protocols::errors::ExecutionError;
+use protocols::frontend::ast::Expr;
+use protocols::frontend::ast::Protocol;
+use protocols::frontend::ast::Stmt;
+use protocols::frontend::ast::StmtId;
 use protocols::frontend::serialize::serialize_bitvec;
+use protocols::frontend::symbol::SymbolId;
+use protocols::frontend::symbol::SymbolTable;
 use rustc_hash::FxHashSet;
+
+use crate::global_context::GlobalContext;
+use crate::interpreter::Interpreter;
+use crate::interpreter::serialize_monitor_args_mapping;
+use crate::queue::*;
+use crate::signal_trace::SignalTrace;
+use crate::signal_trace::WaveSignalTrace;
+use crate::thread::Thread;
+use crate::types::AugmentedProtocolApplication;
+use crate::types::AugmentedTrace;
+use crate::types::CycleResult;
+use crate::types::LoopArgState;
+use crate::types::SchedulerError;
+use crate::types::ThreadResult;
 
 /// Scheduler for handling the multiple threads in the monitor
 #[derive(Clone)]
