+++ monitor/src/thread.rs
@@ -2,12 +2,17 @@
 // released under MIT License
 // author: Ernest Ng <eyn5@cornell.edu>
 
-use crate::{global_context::GlobalContext, types::LoopArgState};
 use baa::BitVecValue;
-use protocols::frontend::ast::{Protocol, StmtId, SymbolId, SymbolTable};
+use protocols::frontend::ast::Protocol;
+use protocols::frontend::ast::StmtId;
 use protocols::frontend::serialize::serialize_stmt;
+use protocols::frontend::symbol::SymbolId;
+use protocols::frontend::symbol::SymbolTable;
 use protocols::scheduler::NextStmtMap;
 use rustc_hash::FxHashMap;
+
+use crate::global_context::GlobalContext;
+use crate::types::LoopArgState;
 
 /// The local context associated with an individual thread,
 /// storing information such as:
