+++ protocols/src/frontend/diagnostic.rs
@@ -9,19 +9,24 @@ use std::io::Write;
 
 use baa::BitVecValue;
 use clap::ColorChoice;
-use codespan_reporting::diagnostic::{
-    Diagnostic as CodespanDiagnostic, Label as CodespanLabel, LabelStyle, Severity,
-};
+use codespan_reporting::diagnostic::Diagnostic as CodespanDiagnostic;
+use codespan_reporting::diagnostic::Label as CodespanLabel;
+use codespan_reporting::diagnostic::LabelStyle;
+use codespan_reporting::diagnostic::Severity;
 use codespan_reporting::files::SimpleFiles;
 use codespan_reporting::term;
-use codespan_reporting::term::termcolor::{Buffer, Color, ColorSpec, WriteColor};
+use codespan_reporting::term::termcolor::Buffer;
+use codespan_reporting::term::termcolor::Color;
+use codespan_reporting::term::termcolor::ColorSpec;
+use codespan_reporting::term::termcolor::WriteColor;
 use pest::RuleType;
 use pest::iterators::Pair;
 use rustc_hash::FxHashSet;
 
 use crate::Value;
 use crate::frontend::ast::*;
-use crate::frontend::serialize::{serialize_bitvec, serialize_value};
+use crate::frontend::serialize::serialize_bitvec;
+use crate::frontend::serialize::serialize_value;
 
 /// Track Errors
 #[derive(Hash, Eq, PartialEq, Debug)]
@@ -549,12 +554,16 @@ impl DiagnosticHandler {
 
 #[cfg(test)]
 mod tests {
+    use std::path::Path;
+
     use baa::BitVecValue;
     use insta::Settings;
-    use std::path::Path;
     use strip_ansi_escapes::strip_str;
 
     use super::*;
+    use crate::frontend::symbol::SymbolKind;
+    use crate::frontend::symbol::SymbolTable;
+    use crate::frontend::symbol::Type;
 
     #[test]
     fn test_emit_diagnostic() {
