+++ protocols/src/frontend/parser.rs
@@ -5,17 +5,20 @@
 // author: Francis Pham <fdp25@cornell.edu>
 // author: Ernest Ng <eyn5@cornell.edu>
 
-use crate::frontend::ast::Stmt;
-use baa::{BitVecValue, WidthInt};
+use std::vec;
+
+use baa::BitVecValue;
+use baa::WidthInt;
 use pest::Parser;
 use pest::error::InputLocation;
 use pest::iterators::Pairs;
 use pest::pratt_parser::PrattParser;
 use pest_derive::Parser;
-use std::vec;
 
+use crate::frontend::ast::Stmt;
 use crate::frontend::ast::*;
 use crate::frontend::diagnostic::*;
+use crate::frontend::symbol::*;
 
 #[derive(Parser)]
 #[grammar = "frontend/protocols.pest"]
