+++ protocols/src/scheduler.rs
@@ -5,20 +5,24 @@
 // author: Francis Pham <fdp25@cornell.edu>
 // author: Ernest Ng <eyn5@cornell.edu>
 
-use baa::{BitVecOps, BitVecValue};
+use baa::BitVecOps;
+use baa::BitVecValue;
 use log::info;
-use rustc_hash::FxHashMap;
-
-use crate::errors::DiagnosticEmitter;
-use crate::errors::{ExecutionError, ExecutionResult};
-use crate::frontend::ast::*;
-use crate::frontend::diagnostic::DiagnosticHandler;
-use crate::interpreter::{Evaluator, ThreadInputValue};
-
-use crate::Value;
 use patronus::expr::Context;
 use patronus::sim::Interpreter;
 use patronus::system::TransitionSystem;
+use rustc_hash::FxHashMap;
+
+use crate::Value;
+use crate::errors::DiagnosticEmitter;
+use crate::errors::ExecutionError;
+use crate::errors::ExecutionResult;
+use crate::frontend::ast::*;
+use crate::frontend::diagnostic::DiagnosticHandler;
+use crate::frontend::symbol::SymbolId;
+use crate::frontend::symbol::SymbolTable;
+use crate::interpreter::Evaluator;
+use crate::interpreter::ThreadInputValue;
 
 /// `NextStmtMap` allows us to interpret without using recursion
 /// (the interpreter can just lookup what the next statement is using this map)
