+++ protocols/src/transactions/mod.rs
@@ -1,7 +1,9 @@
-use crate::Value;
-use crate::frontend::ast;
-use crate::frontend::diagnostic::DiagnosticHandler;
 use rustc_hash::FxHashMap;
+
+use crate::Value;
+use crate::frontend::ast::Protocol;
+use crate::frontend::diagnostic::DiagnosticHandler;
+use crate::frontend::symbol::SymbolTable;
 
 pub mod parser;
 
@@ -10,7 +12,7 @@ pub type Traces = Vec<Vec<(String, Vec<Value>)>>;
 /// Simple frontend for loading a transaction trace file (*.tx)
 pub fn transaction_frontend<'a>(
     filename: impl AsRef<std::path::Path>,
-    protos: impl Iterator<Item = &'a (ast::Protocol, ast::SymbolTable)>,
+    protos: impl Iterator<Item = &'a (Protocol, SymbolTable)>,
     diag: &mut DiagnosticHandler,
 ) -> anyhow::Result<Traces> {
     let protos: FxHashMap<String, _> = protos.map(|(p, st)| (p.name.clone(), (p, st))).collect();
