+++ protocols/src/yosys.rs
@@ -1,5 +1,7 @@
 use std::fs;
-use std::path::{Path, PathBuf};
+use std::path::Path;
+use std::path::PathBuf;
+
 use thiserror::Error;
 
 #[allow(dead_code)]
